requirements.txt
# OpsSentry Requirements
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# API & HTTP
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Logging & Monitoring
colorlog>=6.7.0

# Data Validation
jsonschema>=4.17.0

# Testing
pytest>=7.4.0
pytest-flask>=1.2.0

# Optional: Jupyter for examples
jupyter>=1.0.0
notebook>=6.5.0

# Optional: Advanced ML
# tensorflow>=2.13.0  # Uncomment if using neural networks
# torch>=2.0.0  # Uncomment if using PyTorch
